requirements.txt
obd>=0.7.1
paho-mqtt>=1.5.1
PyYAML>=5.4.1